+++ Welcome/app/src/main/java/com/example/welcome/activity/base/MyBaseActivity.java
@@ -1,0 +1,11 @@
+package com.example.welcome.activity.base;
+
+import android.app.Activity;
+import android.view.View;
+
+/**
+ * Created by HIT-you-CRY on 2015/4/1.
+ */
+public abstract class MyBaseActivity extends Activity implements View.OnClickListener{
+    protected abstract void initView();
+}
