+++ Welcome/app/src/main/res/layout/activity_start.xml
@@ -8,11 +8,12 @@
     android:paddingBottom="@dimen/activity_vertical_margin"
     tools:context=".StartActivity">
 
-    <TextView
-        android:text="@string/hello_world"
-        android:textColor="@color/red"
+    <Button
+        android:id="@+id/button_begin"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
         android:layout_centerInParent="true"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content" />
+        android:text="@string/begin"
+        android:textColor="@color/red"/>
 
 </RelativeLayout>
