+++ Welcome/app/src/main/res/layout/popup_layout.xml
@@ -1,0 +1,23 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:orientation="vertical">
+
+    <Button
+        android:id="@+id/button_wechat_login"
+        android:background="@color/color_ffffcc"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:text="@string/button_wechat_login"
+        android:textColor="@color/red" />
+
+    <Button
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:background="@color/color_ffffcc"
+        android:textColor="@color/red"
+        android:text="@string/button_qq_login"
+        android:id="@+id/button_qq_login" />
+
+</LinearLayout>
