+++ Welcome/app/src/main/res/values/colors.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
     <color name="red">#ff0000</color>
-
+    <color name="color_ffffcc">#ffffcc</color>
 
 </resources>
