+++ Welcome/app/src/main/res/values/strings.xml
@@ -6,5 +6,8 @@
 
     <!-- Start Activity start -->
     <string name="welcome">GitHub</string>
+    <string name="button_wechat_login">WeChat Login</string>
+    <string name="button_qq_login">QQ Login</string>
+    <string name="begin">BEGIN</string>
     <!-- Start Activity end-->
 </resources>
